refactor: 하이라이트 텍스트 기능 컴포넌트(HighlightedText) 분리, 유틸함수화
