+++ src/pages/PostsManagerPage.tsx
@@ -43,6 +43,7 @@ import {
   Textarea,
 } from '../shared/ui';
 import { BaseDialog } from '../widgets/ui/BaseDialog';
+import { HighlightedText } from '../shared/ui/HighlightedText';
 
 import { useSelectedPostStore } from '../stores/posts/useSelectedPostStore';
 import { usePostAddDialog } from '../model/dialog/usePostDialog';
@@ -336,27 +337,6 @@ const PostsManager = () => {
     }
   }, [postsData, usersData, setPosts]);
 
-  // 하이라이트 함수 추가
-  const highlightText = (text: string, highlight: string) => {
-    if (!text) return null;
-    if (!highlight.trim()) {
-      return <span>{text}</span>;
-    }
-    const regex = new RegExp(`(${highlight})`, 'gi');
-    const parts = text.split(regex);
-    return (
-      <span>
-        {parts.map((part, i) =>
-          regex.test(part) ? (
-            <mark key={i}>{part}</mark>
-          ) : (
-            <span key={i}>{part}</span>
-          ),
-        )}
-      </span>
-    );
-  };
-
   // 게시물 테이블 렌더링
   const renderPostTable = () => (
     <Table>
@@ -375,8 +355,9 @@ const PostsManager = () => {
             <TableCell>{post.id}</TableCell>
             <TableCell>
               <div className="space-y-1">
-                <div>{highlightText(post.title, searchQuery)}</div>
-
+                <div>
+                  <HighlightedText text={post.title} highlight={searchQuery} />
+                </div>
                 <div className="flex flex-wrap gap-1">
                   {post.tags?.map((tag) => (
                     <span
@@ -479,7 +460,7 @@ const PostsManager = () => {
                 {comment.user.username}:
               </span>
               <span className="truncate">
-                {highlightText(comment.body, searchQuery)}
+                <HighlightedText text={comment.body} highlight={searchQuery} />
               </span>
             </div>
             <div className="flex items-center space-x-1">
@@ -672,11 +653,21 @@ const PostsManager = () => {
         open={showPostDetailDialog}
         onOpenChange={setShowPostDetailDialog}
         title={
-          selectedPost?.title && highlightText(selectedPost?.title, searchQuery)
+          selectedPost?.title && (
+            <HighlightedText
+              text={selectedPost?.title}
+              highlight={searchQuery}
+            />
+          )
         }
       >
         <p>
-          {selectedPost?.body && highlightText(selectedPost?.body, searchQuery)}
+          {selectedPost?.body && (
+            <HighlightedText
+              text={selectedPost?.body}
+              highlight={searchQuery}
+            />
+          )}
         </p>
         {selectedPost?.id ? renderComments(selectedPost?.id) : null}
       </BaseDialog>
