+++ src/shared/lib/highlightSplitText.ts
@@ -1,0 +1,6 @@
+export const highlightSplitText = (text: string, highlight: string) => {
+  const regex = new RegExp(`(${highlight})`, 'gi');
+  const parts = text.split(regex);
+
+  return { regex, parts };
+};
